modern-cms-platform-pro/backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.2.*
psycopg2-binary==2.9.*
Pillow==10.0.*
python-decouple==3.8.*
gunicorn==21.2.*
whitenoise==6.6.*
django-redis==5.3.*
drf-spectacular==0.27.*
```

#### `backend/.env.example`

```